move document module pipes from shared to document folder
