+++ frontend/src/app/document/pipes/elasticsearch-highlight.pipe.ts
