+++ frontend/src/app/document/pipes/entity.pipe.spec.ts
