+++ frontend/src/app/document/pipes/entity.pipe.ts
@@ -1,7 +1,7 @@
 import { Pipe, PipeTransform } from '@angular/core';
 import { icon } from '@fortawesome/fontawesome-svg-core';
 
-import { entityIcons } from '../icons';
+import { entityIcons } from '../../shared/icons';
 import { FieldEntities } from '../../models';
 
 @Pipe({
