+++ frontend/src/app/document/pipes/geo-data.pipe.ts
