+++ frontend/src/app/document/pipes/paragraph.pipe.ts
