+++ frontend/src/app/document/pipes/snippet.pipe.ts
