+++ frontend/src/app/shared/pipes/index.ts
@@ -1,6 +1,1 @@
-export * from './elasticsearch-highlight.pipe';
-export * from './entity.pipe';
-export * from './geo-data.pipe';
-export * from './paragraph.pipe';
 export * from './regex-highlight.pipe';
-export * from './snippet.pipe';
